+++ file_store.go
@@ -17,10 +17,7 @@ package credentials
 
 import (
 	"context"
-	"encoding/base64"
 	"errors"
-	"fmt"
-	"strings"
 
 	"oras.land/oras-go/v2/registry/remote/auth"
 )
@@ -35,45 +32,9 @@ type FileStore struct {
 	config *config
 }
 
-// configFieldAuths is the "auths" field in the config file.
-// Reference: https://github.com/docker/cli/blob/v24.0.0-beta.1/cli/config/configfile/file.go#L19
-const configFieldAuths = "auths"
-
-var (
-	// ErrInvalidConfigFormat is returned when the config format is invalid.
-	ErrInvalidConfigFormat = errors.New("invalid config format")
-	// ErrPlaintextPutDisabled is returned by Put() when DisablePut is set
-	// to true.
-	ErrPlaintextPutDisabled = errors.New("putting plaintext credentials is disabled")
-)
-
-// newAuthConfig creates an authConfig based on cred.
-func newAuthConfig(cred auth.Credential) authConfig {
-	return authConfig{
-		Auth:          encodeAuth(cred.Username, cred.Password),
-		IdentityToken: cred.RefreshToken,
-		RegistryToken: cred.AccessToken,
-	}
-}
-
-// Credential returns an auth.Credential based on ac.
-func (ac authConfig) Credential() (auth.Credential, error) {
-	cred := auth.Credential{
-		Username:     ac.Username,
-		Password:     ac.Password,
-		RefreshToken: ac.IdentityToken,
-		AccessToken:  ac.RegistryToken,
-	}
-	if ac.Auth != "" {
-		var err error
-		// override username and password
-		cred.Username, cred.Password, err = decodeAuth(ac.Auth)
-		if err != nil {
-			return auth.EmptyCredential, fmt.Errorf("failed to decode auth field: %w: %v", ErrInvalidConfigFormat, err)
-		}
-	}
-	return cred, nil
-}
+// ErrPlaintextPutDisabled is returned by Put() when DisablePut is set
+// to true.
+var ErrPlaintextPutDisabled = errors.New("putting plaintext credentials is disabled")
 
 // NewFileStore creates a new file credentials store.
 func NewFileStore(configPath string) (*FileStore, error) {
@@ -107,30 +68,4 @@ func (fs *FileStore) Put(_ context.Context, serverAddress string, cred auth.Cred
 // Delete removes credentials from the store for the given server address.
 func (fs *FileStore) Delete(_ context.Context, serverAddress string) error {
 	return fs.config.deleteAuthConfig(serverAddress)
-}
-
-// encodeAuth base64-encodes username and password into base64(username:password).
-func encodeAuth(username, password string) string {
-	if username == "" && password == "" {
-		return ""
-	}
-	return base64.StdEncoding.EncodeToString([]byte(username + ":" + password))
-}
-
-// decodeAuth decodes a base64 encoded string and returns username and password.
-func decodeAuth(authStr string) (username string, password string, err error) {
-	if authStr == "" {
-		return "", "", nil
-	}
-
-	decoded, err := base64.StdEncoding.DecodeString(authStr)
-	if err != nil {
-		return "", "", err
-	}
-	decodedStr := string(decoded)
-	username, password, ok := strings.Cut(decodedStr, ":")
-	if !ok {
-		return "", "", fmt.Errorf("auth '%s' does not conform the base64(username:password) format", decodedStr)
-	}
-	return username, password, nil
 }
